+++ todo-frontend/package.json
@@ -16,6 +16,7 @@
     "history": "^5.3.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.5.0",
     "tailwindcss": "^4.1.1"
   },
   "devDependencies": {
